src/backend/requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Environment
python-dotenv==1.0.1

# Google Services
google-adk==0.5.1
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Audio Processing
soundfile==0.12.1
openai-whisper==20231117

# HTTP & Async
httpx==0.27.2
requests==2.32.3

# MCP
fastmcp==0.2.2
mcp==1.1.2

# File Upload
python-multipart==0.0.20